+++ packages/xchain-thorchain-query/__e2e__/thorchain-estimateSwap.e2e.ts
@@ -34,6 +34,7 @@ function print(estimate: SwapEstimate, input: CryptoAmount) {
       swapFee: estimate.totalFees.swapFee.formatedAssetString(),
       outboundFee: estimate.totalFees.outboundFee.formatedAssetString(),
       affiliateFee: estimate.totalFees.affiliateFee.formatedAssetString(),
+      // totalFees: estimate.totalFees.totalFees?.formatedAssetString(),
     },
     slipPercentage: estimate.slipPercentage.toFixed(),
     netOutput: estimate.netOutput.formatedAssetString(),
@@ -54,16 +55,31 @@ function printTx(txDetails: TxDetails, input: CryptoAmount) {
 }
 const BUSD = assetFromStringEx('BNB.BUSD-BD1')
 const BTCB = assetFromStringEx('BNB.BTCB-1DE')
+const USDCETH = assetFromStringEx('ETH.USDC-0XA0B86991C6218B36C1D19D4A2E9EB0CE3606EB48')
 
 // Test User Functions - single and double swap using mock pool data
 describe('Thorchain-query estimate Integration Tests', () => {
   // Test estimate swaps with mock pool data
+  it('should estimate a swap of 10 BTC to RUNE with 0.3 affiliate fee', async () => {
+    const swapParams: EstimateSwapParams = {
+      input: new CryptoAmount(assetToBase(assetAmount('10')), AssetBTC),
+      destinationAsset: AssetRuneNative,
+      destinationAddress: 'xxx',
+      affiliateAddress: 'affiliateAddress',
+      affiliateFeeBasisPoints: 30, //optional
+    }
+
+    const estimate = await thorchainQuery.estimateSwap(swapParams)
+    // const estimateInBusd = await thorchainQuery.getFeesIn(estimate.txEstimate.totalFees, BUSD)
+    // estimate.txEstimate.totalFees = estimateInBusd
+    printTx(estimate, swapParams.input)
+  })
   it('should estimate a swap of 0.5 BTC to BUSD', async () => {
     const swapParams: EstimateSwapParams = {
       input: new CryptoAmount(assetToBase(assetAmount('0.5')), AssetBTC),
       destinationAsset: BUSD,
       destinationAddress: 'xxx',
-      // affiliateFeePercent: 0.003, //optional
+      // affiliateFeeBasisPoints: 0.003, //optional
       slipLimit: new BigNumber('0.03'), //optional
     }
 
@@ -85,7 +101,7 @@ describe('Thorchain-query estimate Integration Tests', () => {
       input: new CryptoAmount(assetToBase(assetAmount('0.5')), BTCB),
       destinationAsset: BUSD,
       destinationAddress: 'xxx',
-      // affiliateFeePercent: 0.003, //optional
+      // affiliateFeeBasisPoints: 0.003, //optional
       slipLimit: new BigNumber('0.03'), //optional
     }
 
@@ -112,7 +128,7 @@ describe('Thorchain-query estimate Integration Tests', () => {
       input: new CryptoAmount(assetToBase(assetAmount(1)), AssetBTC),
       destinationAsset: sBTC,
       destinationAddress: 'xxx',
-      // affiliateFeePercent: 0.03, //optional
+      // affiliateFeeBasisPoints: 0.03, //optional
       slipLimit: new BigNumber(0.03), //optional
     }
     const estimate = await thorchainQuery.estimateSwap(swapParams)
@@ -129,7 +145,7 @@ describe('Thorchain-query estimate Integration Tests', () => {
       input: new CryptoAmount(assetToBase(assetAmount(1)), sBTC),
       destinationAsset: sETH,
       destinationAddress: 'xxx',
-      affiliateFeePercent: 0.03, //optional
+      affiliateFeeBasisPoints: 30, //optional
       slipLimit: new BigNumber(0.02), //optional
     }
     const estimate = await thorchainQuery.estimateSwap(swapParams)
@@ -204,7 +220,7 @@ describe('Thorchain-query estimate Integration Tests', () => {
     const swapParams: EstimateSwapParams = {
       input: new CryptoAmount(assetToBase(assetAmount(1)), AssetBTC),
       destinationAsset: AssetETH,
-      affiliateFeePercent: 0 || 1001,
+      affiliateFeeBasisPoints: 0 || 1001,
       destinationAddress: 'xxx',
     }
     try {
@@ -212,7 +228,7 @@ describe('Thorchain-query estimate Integration Tests', () => {
       print(estimate.txEstimate, swapParams.input)
       fail()
     } catch (error: any) {
-      expect(error.message).toEqual(`affiliateFee must be between 0 and 1000`)
+      expect(error.message).toEqual(`affiliateFeeBasisPoints must be between 0 and 1000 basis points`)
     }
   })
 
@@ -221,7 +237,7 @@ describe('Thorchain-query estimate Integration Tests', () => {
       input: new CryptoAmount(assetToBase(assetAmount(50)), AssetBTC),
       destinationAsset: AssetETH,
       destinationAddress: 'xxx',
-      affiliateFeePercent: 0.03, //optional
+      affiliateFeeBasisPoints: 30, //optional
       slipLimit: new BigNumber(0.02), //optional
     }
     try {
@@ -240,13 +256,13 @@ describe('Thorchain-query estimate Integration Tests', () => {
       synth: false,
     }
     const swapParams: EstimateSwapParams = {
-      input: new CryptoAmount(assetToBase(assetAmount(20)), AssetETH),
+      input: new CryptoAmount(assetToBase(assetAmount(20, 18)), AssetETH),
       destinationAsset: assetHOT,
       destinationAddress: 'xxx',
     }
     try {
       const estimate = await thorchainQuery.estimateSwap(swapParams)
-      expect(estimate.txEstimate.errors).toEqual([`destinationAsset HOT does not have a valid liquidity pool`])
+      expect(estimate.txEstimate.errors[0]).toEqual(`destinationAsset HOT does not have a valid liquidity pool`)
     } catch (error) {
       throw error
     }
@@ -256,7 +272,7 @@ describe('Thorchain-query estimate Integration Tests', () => {
     const swapParams: EstimateSwapParams = {
       input: new CryptoAmount(assetToBase(assetAmount(0.001)), AssetBTC),
       destinationAsset: AssetETH,
-      affiliateFeePercent: 0.1,
+      affiliateFeeBasisPoints: 10,
       destinationAddress: 'xxx',
     }
     try {
@@ -288,7 +304,7 @@ describe('Thorchain-query estimate Integration Tests', () => {
       input: new CryptoAmount(assetToBase(assetAmount('20')), AssetRuneNative),
       destinationAsset: AssetAVAX,
       destinationAddress: 'xxx',
-      // affiliateFeePercent: 0.003, //optional
+      // affiliateFeeBasisPoints: 0.003, //optional
       slipLimit: new BigNumber('0.20'), //optional
     }
 
@@ -339,6 +355,8 @@ describe('Thorchain-query estimate Integration Tests', () => {
       input: new CryptoAmount(assetToBase(assetAmount('5', 8)), AssetBTC),
       destinationAsset: AssetETH,
       destinationAddress: 'xxx',
+      affiliateAddress: `tthor13q9z22fvjkk8r8sxf7hmp2t56jyvn9s7sxx8lx`,
+      affiliateFeeBasisPoints: 30,
       slipLimit: new BigNumber('0.2'),
     }
     const estimate = await thorchainQuery.estimateSwap(swapParams)
@@ -347,15 +365,18 @@ describe('Thorchain-query estimate Integration Tests', () => {
     expect(estimate).toBeTruthy()
   })
 
-  it(`Should estimate a swap from BTC to ETH`, async () => {
-    const swapParams = {
-      input: new CryptoAmount(assetToBase(assetAmount('5', 8)), AssetBTC),
-      destinationAsset: AssetETH,
+  it(`Should estimate a swap from BTC to ETHUSDC and check memo`, async () => {
+    const swapParams: EstimateSwapParams = {
+      input: new CryptoAmount(assetToBase(assetAmount('1', 8)), AssetBTC),
+      destinationAsset: USDCETH,
       destinationAddress: 'xxx',
+      interfaceID: '999',
+      affiliateAddress: `tthor13q9z22fvjkk8r8sxf7hmp2t56jyvn9s7sxx8lx`,
+      affiliateFeeBasisPoints: 30,
       slipLimit: new BigNumber('0.2'),
     }
     const estimate = await thorchainQuery.estimateSwap(swapParams)
-    print(estimate.txEstimate, swapParams.input)
+    printTx(estimate, swapParams.input)
     expect(estimate.txEstimate.canSwap).toBe(true)
     expect(estimate).toBeTruthy()
   })