+++ packages/xchain-thorchain-query/CHANGELOG.md
@@ -1,10 +1,25 @@
+# v0.1.5 (2022-10-27)
+
+## Fix
+
+- Bug - Limit asset amount was not using 8 decimal places
+- Bug - Swap to Synths was failing Pool Conversion 
+- 
+
+## Update
+
+- Removed getInboundAddresses() from thorchainCache since it was redundant
+- Renamed AffiliateFeePercent to AffiliateFeeBasisPoints in EstimateSwapParams
+- Updated output decimals to us NativeDecimals from Pool data
+- Created calcOutboundFee() to use thornode outbound_fee 
+
 # v0.1.4 (2022-10-17)
 
 ## Fix
 
-- Bug - Limit asset amount was not using 8 decimal places 
+- Bug - Limit asset amount was not using 8 decimal places
 - Bug - outbound fee is now calculated correctly
-- Bug - affiliate fee calculation fixed 
+- Bug - affiliate fee calculation fixed
 - Bug - getPoolForAsset() changed to take both chain & ticker for lookup
 - Bump dep on package `xchain-midgard`